added button to switch between sms and email. refactored
