+++ src/assets/App.css
@@ -1,0 +1,56 @@
+/* src/App.css */
+
+body {
+  margin: 0;
+  padding: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  background-color: #0a0a23; /* Very dark blue background */
+  color: #ffffff; /* Default font color white */
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+}
+
+.App {
+  text-align: center;
+}
+
+.App-header {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+button {
+  margin: 10px;
+  padding: 10px 20px;
+  font-size: 18px;
+  cursor: pointer;
+  background-color: #4b0082; /* Indigo background for buttons */
+  color: #ffffff; /* White text on buttons */
+  border: none;
+  border-radius: 5px;
+}
+
+button:hover {
+  background-color: #61dafb; /* Lighter blue on hover */
+  color: #4b0082; /* Indigo text on hover */
+}
+
+h2 {
+  color: #9b30ff; /* Purple color for headings */
+}
+
+p {
+  font-size: 20px;
+  color: #ffffff; /* White text for paragraphs */
+}
