+++ src/assets/landing.css
@@ -1,0 +1,35 @@
+/* src/components/Instructions.css */
+
+.toggle-container {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 20px;
+}
+
+.toggle-option {
+  padding: 10px 20px;
+  cursor: pointer;
+  background-color: #4b0082; /* Default indigo background */
+  color: white; /* White text */
+  border: 1px solid #4b0082; /* Border matching the background */
+  border-radius: 5px 5px 0 0;
+  font-size: 18px;
+  transition: background-color 0.3s, color 0.3s;
+}
+
+.toggle-option.active {
+  background-color: #61dafb; /* Lighter blue when active */
+  color: #4b0082; /* Indigo text when active */
+}
+
+.toggle-option:first-child {
+  border-radius: 5px 0 0 5px;
+}
+
+.toggle-option:last-child {
+  border-radius: 0 5px 5px 0;
+}
+
+.toggle-option:not(:last-child) {
+  border-right: none;
+}
