+++ src/components/email.js
@@ -1,0 +1,14 @@
+// src/components/EmailInstructions.js
+
+import React from 'react';
+
+const EmailInstructions = () => {
+  return (
+    <div>
+      <h2>Email Instructions</h2>
+      <p>Here are the instructions on how to use the email feature...</p>
+    </div>
+  );
+};
+
+export default EmailInstructions;
