+++ src/components/sms.js
@@ -1,0 +1,14 @@
+// src/components/SmsInstructions.js
+
+import React from 'react';
+
+const SmsInstructions = () => {
+  return (
+    <div>
+      <h2>SMS Instructions</h2>
+      <p>Here are the instructions on how to use the SMS feature...</p>
+    </div>
+  );
+};
+
+export default SmsInstructions;
